pve-common: Substitute diff and ip
We already propagate diff and ip binaries via propagatedBuildInputs
but this does not seems to works everywhere.

Take the iface_set_master subroutine in pve-common/src/PVE/Network.pm

I thought it would know how to find ip but it doesn't for some
reason. Is it because it is in an eval block and the context is
somewhat different?