lxc: Add lxc package
This was taken straight out of the nixpkgs definition, there is
probably a better way than doing this.

This was done because there were hardcoded stuff in the lxc package
which I needed to replace and it was easier to refer to this package
from pve-container.

I only made 2 modifications compared to the nixpkgs all in the
postInstall section:

I add this at the end, because it was complaining about missing coreutils:

    sed -i $out/libexec/lxc/lxc-containers \
      -e "s|touch|${coreutils}/bin/touch|" \
      -e "s|rm|${coreutils}/bin/rm|"

And I removed this line:

     substituteInPlace $out/share/lxc/config/common.conf --replace-fail "$out/share" "/run/current-system/sw/share"

/run/current-system/sw/share was empty, not sure how it's supposed to
be populated.