container.nix: Add lxc and lxc-net services
Added these services because that's how it's done in a vanilla proxmox
install, but I think pve-container service being presents are enough
IIRC, we could try removing the wants lxc.

Also instead of defining these here, maybe we could leverage the
builtin virtualisation.lxc? I did not do so because it was
quicker/easier with these services present directly.