pve-container: Fixed lxc includes
LXC config file that gets generated needs to include a lot of stuff.

First is LXCFS so that containers actually know what are their true
uptime for example not the host.

Second seccomp rules

And the rest are common config, note that ideally it also wants distro
specific files, these are not packaged for nix yet.

https://github.com/lxc/lxc-templates/tree/main/config